refactor: clean up the code
