+++ .gitignore
@@ -24,3 +24,4 @@ yarn-error.log*
 .idea
 dist
 .git
+/src/react-app-env.d.ts
