+++ src/app/components/select/_select.scss
@@ -5,6 +5,7 @@ select option:disabled {
 select:invalid{
   color: gray;
 }
+
 option{
   color: black;
   padding-right: 1rem;
